Created category package that has Category -entity, repository, controller and controller service fully implemented and tested. Category added to Task -entity, removed category -enum attribute. Created many-to-many relationship between task and Category using annotations. Created many-to-many relationship between task and location using annotations. Updated all classe example testdata -classes for updated Task -entity. Uptaded TaskController and service that now you can filter user tasks and all tasks with category-title and using same service-method. All tested passed and added couple integrationtests for Task.

+++ backend/glig/src/test/java/com/jttam/glig/domain/location/LocationServiceTest.java
@@ -30,7 +30,7 @@ class LocationServiceTest {
 
     private Location location;
     private LocationResponse locationResponse;
-    private static final Integer LOCATION_ID = 1;
+    private static final Long LOCATION_ID = 1L;
 
     @BeforeEach
     void setUp() {
@@ -40,13 +40,13 @@ class LocationServiceTest {
         location.setCity("Helsinki");
 
         locationResponse = new LocationResponse(
-            LOCATION_ID,        // locationId
-            null,               // streetAddress
-            null,               // postalCode
-            "Helsinki",         // city
-            null,               // country
-            null,               // latitude
-            null                // longitude
+                LOCATION_ID, // locationId
+                null, // streetAddress
+                null, // postalCode
+                "Helsinki", // city
+                null, // country
+                null, // latitude
+                null // longitude
         );
     }
 
@@ -71,10 +71,12 @@ class LocationServiceTest {
 
     @Test
     void getLocation_shouldThrowNotFoundException_whenNotFound() {
-        // Given: The repository will return an empty optional, simulating a not-found scenario
+        // Given: The repository will return an empty optional, simulating a not-found
+        // scenario
         when(locationRepository.findById(LOCATION_ID)).thenReturn(Optional.empty());
 
-        // When & Then: We assert that calling the service method throws the expected exception
+        // When & Then: We assert that calling the service method throws the expected
+        // exception
         assertThrows(NotFoundException.class, () -> {
             locationService.getLocation(LOCATION_ID);
         });
